Fix file deletion issue

+++ src/other.ts
@@ -24,9 +24,11 @@ export function clearV1 (): Empty {
     if (err) throw err;
 
     for (const file of files) {
-      fs.unlink(path.join(directory, file), (err) => {
-        if (err) throw err;
-      });
+      if (file !== '.gitkeep') {
+        fs.unlink(path.join(directory, file), (err) => {
+          if (err) throw err;
+        });
+      }
     }
   });
   return {};
